Small commit.
Now all material objects can have an unlimited form, if the coordinates are filled in a clockwise direction.
The collision algorithms are completely redesigned and adapted to work with arbitrary polygons. Transferred to the folder of material objects. Added additional collision methods.
A Block base class has been added. A collection of blocks is added to the unifier.
Redesigned possibilities.
Some mathematical computations are highlighted in a separate module and moved in the usage pest.